requirements.txt
pandas~=2.3.3
flask~=3.1.2
python-dotenv~=1.2.1
opencv-python~=4.12.0.88
pycryptodome~=3.23.0
paddleocr~=3.3.0
openai-whisper~=20250625
future~=1.0.0
requests~=2.32.5
selenium~=4.38.0
torch~=2.9.0
numpy~=2.2.6
pydub~=0.25.1
setuptools~=80.9.0
ollama~=0.6.0
werkzeug~=3.1.3
openai~=2.7.1
lxml~=6.0.2
rsa~=4.9.1
pyexecjs~=1.5.1
pillow~=12.0.0
uvicorn~=0.38.0
fastapi~=0.121.1
pydantic~=2.12.3
openpyxl~=3.1.5
jieba~=0.42.1
chardet~=5.2.0
urllib3~=2.5.0
bs4~=0.0.2
beautifulsoup4~=4.14.2
playwright~=1.56.0
webencodings~=0.5.1
tqdm~=4.67.1
util~=0.0.1
webdriver-manager~=4.0.2
matplotlib~=3.10.7
wordcloud~=1.9.4
seqeval~=1.2.2
paddlenlp~=2.6.1